# [fix](date) return right date value even if out of the range of date dictionary