--- a/be/src/vec/exec/format/orc/vorc_reader.h
+++ b/be/src/vec/exec/format/orc/vorc_reader.h
@@ -406,7 +406,7 @@ private:
         if (data == nullptr) {
             return Status::InternalError("Wrong data type for colum '{}'", col_name);
         }
-        auto* __restrict date_day_offset_dict = get_date_day_offset_dict();
+        date_day_offset_dict& date_dict = date_day_offset_dict::get();
         auto& column_data = static_cast<ColumnVector<DorisColumnType>&>(*data_column).get_data();
         auto origin_size = column_data.size();
         column_data.resize(origin_size + num_values);
@@ -423,14 +423,12 @@ private:
                     }
                 }
                 int64_t date_value = data->data[i] + _offset_days;
-                DCHECK_LT(date_value, 25500);
-                DCHECK_GE(date_value, 0);
                 if constexpr (std::is_same_v<CppType, VecDateTimeValue>) {
-                    v.create_from_date_v2(date_day_offset_dict[date_value], TIME_DATE);
+                    v.create_from_date_v2(date_dict[date_value], TIME_DATE);
                     // we should cast to date if using date v1.
                     v.cast_to_date();
                 } else {
-                    v = date_day_offset_dict[date_value];
+                    v = date_dict[date_value];
                 }
             } else { // timestamp
                 if constexpr (is_filter) {