--- a/be/src/vec/exec/format/parquet/fix_length_plain_decoder.cpp
+++ b/be/src/vec/exec/format/parquet/fix_length_plain_decoder.cpp
@@ -262,7 +262,7 @@ Status FixLengthPlainDecoder::_decode_date(MutableColumnPtr& doris_column,
     size_t data_index = column_data.size();
     column_data.resize(data_index + select_vector.num_values() - select_vector.num_filtered());
     ColumnSelectVector::DataReadType read_type;
-    auto* __restrict date_day_offset_dict = get_date_day_offset_dict();
+    date_day_offset_dict& date_dict = date_day_offset_dict::get();
 
     while (size_t run_length = select_vector.get_next_run<has_filter>(&read_type)) {
         switch (read_type) {
@@ -271,16 +271,13 @@ Status FixLengthPlainDecoder::_decode_date(MutableColumnPtr& doris_column,
                 char* buf_start = _data->data + _offset;
                 int64_t date_value = static_cast<int64_t>(*reinterpret_cast<int32_t*>(buf_start)) +
                                      _decode_params->offset_days;
-                DCHECK_LT(date_value, 25500);
-                DCHECK_GE(date_value, 0);
                 if constexpr (std::is_same_v<CppType, VecDateTimeValue>) {
                     auto& v = reinterpret_cast<CppType&>(column_data[data_index++]);
-                    v.create_from_date_v2(date_day_offset_dict[date_value], TIME_DATE);
+                    v.create_from_date_v2(date_dict[date_value], TIME_DATE);
                     // we should cast to date if using date v1.
                     v.cast_to_date();
                 } else {
-                    reinterpret_cast<CppType&>(column_data[data_index++]) =
-                            date_day_offset_dict[date_value];
+                    reinterpret_cast<CppType&>(column_data[data_index++]) = date_dict[date_value];
                 }
                 _offset += _type_length;
             }