… regression-test/data/external_table_p2/hive/test_complex_types.out