## Proposed changes

PR(https://github.com/apache/doris/pull/22360) and PR(https://github.com/apache/doris/pull/22384) optimized the performance of date type. However hive supports date out of 1970~2038, leading wrong date value in tpcds benchmark.
How to fix:
1. Increase dictionary range: 1900 ~ 2038
2. The date out of 1900 ~ 2038 is regenerated.

## Performance
Single thread to scan 512000000 rows:
| date range | \<1900 | 1900 ~ 2038 | \> 2038 |
|  ----  | ----  | ---- | ---- |
| previous PR. | - | 11.92s | - |
| this PR. |  37.87s | 12.73s | 40.67s |

## Remaining problems
Date like `0002-11-30` will be generated wrongly.

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...